Create loan applications for users
A LoanApplication holds the questionnaire data and will be used by the
loan product wrapper to filter potential loans.

There's no unit testing for this model as it is only providing storage
at this point, with no logic. I was tempted to write at least a simple
assertion that the model could be created appropriately, but decided
against it.