Wrap LH Products API and filter results
The LhProductsApiWrapper makes the GET request to the API, pulls down
the hash of available products and then filters it against a provide
LoanApplication instance. If the GET is unsuccessful or their are no
matching products then an empty array is returned so that consumers of
the wrapper can rely on working with an Enumerable.

I was tempted to write a LH Product PORO to hold the results and then
make working with them in the views cleaner, rather than working with a
raw hash. However, the exercise was taking longer than expected, so I
passed and stuck with the hash results.