Move to RESTful routes for Users and Loans
I prefer to default to RESTful routes unless there is a compelling
reason not to. The aesthetics of the routes can be achieved by using
named routes, but it requires passing variables as locals to views.

I removed the ActiveRecord::Base.transaction call since the
accepts_nested_attributes_for on User handles this implicitly now.
I added strong params as well.

I'm not entirely pleased with the workflow in the UsersController. I
think it would be better if the loan selection step was on the loans
controller via a #create action. However, I had that feeling after
completing the functionality and did not feel the change warranted
further time.