User to accept nested attributes & Loan cleanup
We no longer need the hard coded values in the Loan model, since those
will be set by the values passed from the view.

By accepting nested attributes for the associated loan and
loan_application, the modifications to the User instance and associate
stays atomic without needing to explicitly wrap modifications in a
transaction.