Create views for loan application, selection and confirmation
I replaced the main/home view with users/new view which has fields_for
the nested LoanApplication.

The next screen is the users/edit view which uses two partials
('no_options', 'loan_options') depending on what is in the @products
instance variable. I'm not a huge fan of having logic in a view. Another
alternative would be to push that logic into the controller to select
the view, which isn't ideal. The best would to move the logic to
ViewObject/Presenter.

I replaced the main/thank_you with loans/show, since the data required
for the view is the loan.name.

The styling could be a lot better. This was my first time working with
Bulma, which was pleasant enough. Just not something I was familiar
with. As you can see the loan selection screen is far cleaner looking
after I had taken a few minutes to read through the docs and find some
useful classes to apply. I did look at the docs for styling the dropdown
list and decided against pursuing that further.

I'm also not happy with manipulating the product results as a hash. It's
sloppy looking and instances where some logic was applied, as in the
case of the term converting from months to years, is done right in the
view which is poor form.

The creation of the buttons is a bit clumsy as well and would be better
in a helper; again having the product encapsulated in a class would
clean that eventual code up as well.

The Carmen gem comes into play as well for the loan_application form.
While US States are basically set in stone, the addition was simple and
clean while providing for a clean expansion point if geographic
requirements become more complex.